requirements.txt
# MentorIQ Backend Dependencies for Production
fastapi==0.100.1
uvicorn==0.23.2
groq==0.30.0
httpx==0.24.1
python-dotenv==1.0.0
pydantic==2.11.7